requirements.txt
# NZB Indexer Requirements
# Version: 0.8.0

# FastAPI and web server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
starlette>=0.41.0
jinja2>=3.1.4
python-multipart>=0.0.12
itsdangerous>=2.2.0  # For session middleware

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0  # PostgreSQL driver
aiosqlite>=0.20.0  # SQLite driver

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.10.0
pydantic-settings>=2.6.1
python-dotenv>=1.0.1

# Utilities
email-validator>=2.2.0
httpx>=0.28.0
tenacity>=9.0.0
